requirements.txt
scipion-em-tomo
scipion-em-tomo3D
scipion-em-relion
scipion-em-reliontomo